Move accumulation functions out of the long_range_adjacency module
