+++ src/python/lib/graph/long_range_adjacency/CMakeLists.txt
@@ -7,7 +7,6 @@ addPythonModule(
         adjacency.cxx 
         long_range_adjacency.cxx
         long_range_features.cxx
-        accumulate_long_range_affinities.cxx
     LIBRRARIES
         ${HDF5_LIBRARIES}
         ${Boost_SYSTEM_LIBRARY}
