+++ src/python/module/nifty/graph/__init__.py
@@ -247,3 +247,100 @@ def run_label_propagation(graph, edge_values=None, nb_iter=1, local_edges=None, 
     runLabelPropagation_impl(graph, node_labels, edge_values, local_edges, nb_iter, size_constr, nb_threads)
 
     return node_labels
+
+
+import numpy as np
+import nifty.graph.rag as nrag
+
+def accumulate_affinities_mean_and_length(affinities, offsets, labels, graph=None,
+                                          affinities_weights=None,
+                                          offset_weights=None,
+                                          ignore_label=None, number_of_threads=-1):
+    """
+    Features of this function (additional ones compared to other accumulate functions):
+      - does not require a RAG but simply a graph and a label image (can include long-range edges)
+      - can perform weighted average of affinities depending on given affinitiesWeights
+      - ignore pixels with ignore label
+
+    Parameters
+    ----------
+    affinities: offset channels expected to be the first one
+    """
+    affinities = np.require(affinities, dtype='float32')
+
+    if affinities_weights is not None:
+        assert offset_weights is None, "Affinities weights and offset weights cannot be passed at the same time"
+        affinities_weights = np.require(affinities_weights, dtype='float32')
+
+    else:
+        affinities_weights = np.ones_like(affinities)
+        if offset_weights is not None:
+            offset_weights = np.require(offset_weights, dtype='float32')
+            for _ in range(affinities_weights.ndim-1):
+                offset_weights = np.expand_dims(offset_weights, axis=-1)
+            affinities_weights *= offset_weights
+
+    affinities = np.rollaxis(affinities, axis=0, start=len(affinities.shape))
+    affinities_weights = np.rollaxis(affinities_weights, axis=0, start=len(affinities_weights.shape))
+
+    offsets = np.require(offsets, dtype='int32')
+    assert len(offsets.shape) == 2
+
+    if graph is None:
+        graph = nrag.gridRag(labels)
+
+
+    hasIgnoreLabel = (ignore_label is not None)
+    ignore_label = 0 if ignore_label is None else int(ignore_label)
+
+    number_of_threads = -1 if number_of_threads is None else number_of_threads
+
+    edge_indicators_mean, edge_indicators_max, edge_sizes = \
+        accumulateAffinitiesMeanAndLength_impl_(
+            graph,
+            labels.astype('uint64'),
+            affinities,
+            affinities_weights,
+            offsets,
+            hasIgnoreLabel,
+            ignore_label,
+            number_of_threads
+        )
+    return edge_indicators_mean, edge_sizes
+
+
+def accumulate_affinities_mean_and_length_inside_clusters(affinities, offsets, labels,
+                                                          offset_weights=None,
+                                                          ignore_label=None, number_of_threads=-1):
+    """
+    Similar idea to `accumulate_affinities_mean_and_length`, but accumulates affinities/edge-values for all edges not on
+     cut (i.e. connecting nodes in the same cluster)
+    """
+    affinities = np.require(affinities, dtype='float32')
+    affinities = np.rollaxis(affinities, axis=0, start=len(affinities.shape))
+
+    offsets = np.require(offsets, dtype='int32')
+    assert len(offsets.shape) == 2
+
+    if offset_weights is None:
+        offset_weights = np.ones(offsets.shape[0], dtype='float32')
+    else:
+        offset_weights = np.require(offset_weights, dtype='float32')
+
+    hasIgnoreLabel = (ignore_label is not None)
+    ignore_label = 0 if ignore_label is None else int(ignore_label)
+
+    number_of_threads = -1 if number_of_threads is None else number_of_threads
+
+    edge_indicators_mean, edge_indicators_max, edge_sizes = \
+        accumulateAffinitiesMeanAndLengthInsideClusters_impl_(
+            labels.astype('uint64'),
+            labels.max(),
+            affinities,
+            offsets,
+            offset_weights,
+            hasIgnoreLabel,
+            ignore_label,
+            number_of_threads
+        )
+    return edge_indicators_mean, edge_sizes
